+++ tests/installation/connect-to-selenium.spec.ts
@@ -16,8 +16,8 @@
 import path from 'path';
 import { test } from './npmTest';
 
-test('connect to selenium', async ({ exec, tmpWorkspace }) => {
+test('connect to selenium', async ({ exec, tmpWorkspace }, testInfo) => {
   await exec('npm i --foreground-scripts playwright-core');
   await exec(`node ./download-chromedriver.js ${path.join(tmpWorkspace)}`);
-  await exec('npm run test -- --reporter=list selenium.spec', { cwd: path.join(__dirname, '..', '..'), env: { ...process.env, PWTEST_CHROMEDRIVER: path.join(tmpWorkspace, 'chromedriver') } });
+  await exec(`npm run test -- --reporter=list selenium.spec --output=${testInfo.outputPath('tmp-test-results')}`, { cwd: path.join(__dirname, '..', '..'), env: { PWTEST_CHROMEDRIVER: path.join(tmpWorkspace, 'chromedriver') } });
 });
