+++ tests/installation/expect.d.ts
@@ -3,7 +3,7 @@ export {}
 declare global {
     namespace PlaywrightTest {
        interface Matchers<R, T> {
-          toHaveDownloaded(browsers: ("chromium" | "firefox" | "webkit")[]): R;
+          toHaveLoggedSoftwareDownload(browsers: ("chromium" | "firefox" | "webkit" | "ffmpeg")[]): R;
        }
     }
 }
