+++ tests/installation/npx-global-help.spec.ts
@@ -15,9 +15,9 @@
  */
 import { test, expect } from './npmTest';
 
-test('npx playwright --help should not download browsers', async ({ exec, installedBrowsers }) => {
+test('npx playwright --help should not download browsers', async ({ exec, installedSoftwareOnDisk }) => {
   const result = await exec('npx playwright --help');
-  expect(result).toHaveDownloaded([]);
-  expect(await installedBrowsers()).toEqual([]);
+  expect(result).toHaveLoggedSoftwareDownload([]);
+  expect(await installedSoftwareOnDisk()).toEqual([]);
   expect(result).not.toContain(`To avoid unexpected behavior, please install your dependencies first`);
 });
