+++ tests/installation/playwright-global-installation.spec.ts
@@ -15,9 +15,10 @@
  */
 import { test, expect } from './npmTest';
 
-test('global installation', async ({ exec, installedBrowsers }) => {
+test('global installation', async ({ exec, installedSoftwareOnDisk }) => {
   const result = await exec('npm i --foreground-scripts playwright');
-  (expect(result) as any).toHaveDownloaded(['chromium', 'firefox', 'webkit']);
-  expect(await installedBrowsers()).toEqual(['chromium', 'firefox', 'webkit']);
+  expect(result).toHaveLoggedSoftwareDownload(['chromium', 'firefox', 'webkit']);
+  expect(await installedSoftwareOnDisk()).toEqual(['chromium', 'firefox', 'webkit']);
+  await exec('node sanity.js playwright none', { env: { PLAYWRIGHT_BROWSERS_PATH: undefined } });
   await exec('node ./sanity.js playwright');
 });
