+++ tests/installation/playwright-test-stacks-should-work.spec.ts
@@ -17,8 +17,8 @@ import { test, expect } from './npmTest';
 
 test('@playwright/test stacks should work', async ({ exec }) => {
   await exec('npm i --foreground-scripts @playwright/test');
-  await exec('npx playwright install chromium', { env: { PLAYWRIGHT_BROWSERS_PATH: '0' } });
-  const output = await exec('npx playwright test -c . failing.spec.js', { expectToExitWithError: true, env: { DEBUG: 'pw:api', PLAYWRIGHT_BROWSERS_PATH: '0' } });
+  await exec('npx playwright install chromium');
+  const output = await exec('npx playwright test -c . failing.spec.js', { expectToExitWithError: true, env: { DEBUG: 'pw:api' } });
   expect(output).toContain('expect.toHaveText started');
   expect(output).toContain('failing.spec.js:5:38');
 });
