+++ tests/installation/skip-browser-download-inspect-with-custom-executable.spec.ts
@@ -16,11 +16,10 @@
 import { test, expect } from './npmTest';
 
 test('should skip download', async ({ exec }) => {
-  const env = { PLAYWRIGHT_SKIP_BROWSER_DOWNLOAD: '1' };
-  const installOutput = await exec('npm i --foreground-scripts playwright', { env });
+  const installOutput = await exec('npm i --foreground-scripts playwright', { env: { PLAYWRIGHT_SKIP_BROWSER_DOWNLOAD: '1' } });
   expect(installOutput).toContain('Skipping browsers download because');
   if (process.platform === 'linux') {
-    const output = await exec('node inspector-custom-executable.js', { env });
+    const output = await exec('node inspector-custom-executable.js', { env: { PWDEBUG: '1' } });
     expect(output).toContain('SUCCESS');
   }
 });
